requirements.txt
django
Pillow
pandas
plotly
networkx
os
openpyxl